pyproject.toml
[tool.poetry]
name = "pytorch-summary"
version = "0.1.0"
description = "Pytorch summary (based on sksq96/pytorch-summary) which supports arbitrary inputs"
authors = ["Venky Iyer <indigoviolet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = ">=5.4.3"
torch = ">=1.5.1"
attrs = ">=19.3.0"
tabulate = ">=0.8.7"
humanfriendly = ">=8.2"
shapely = {git = "https://github.com/indigoviolet/shapely.git"}

[tool.poetry.dev-dependencies]

[tool.dephell.main]
# See https://github.com/python-poetry/poetry/issues/761#issuecomment-521124268
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
